compiler: Reorganize
Move functions implementing compilation of the query AST to methods of
the Compiler class. This allows to more easily pass the query
compilation context between different functions. Only code
reorganization, no intended functional changes.
