+++ beanquery/query_compile.py
@@ -8,7 +8,6 @@ __copyright__ = "Copyright (C) 2014-2016  Martin Blais"
 __license__ = "GNU GPLv2"
 
 import collections
-import copy
 import datetime
 import re
 import operator
@@ -16,28 +15,14 @@ import operator
 from decimal import Decimal
 from itertools import product
 
-from beanquery import parser
+from beanquery.parser import ast
 from beanquery import query_execute
 from beanquery import types
 from beanquery import tables
-from beanquery.parser import ast
-
-
-# A global constant which sets whether we support inferred/implicit group-by
-# semantics.
-SUPPORT_IMPLICIT_GROUPBY = True
 
 
 FUNCTIONS = collections.defaultdict(list)
 OPERATORS = collections.defaultdict(list)
-
-
-class CompilationError(Exception):
-    """A compiler/interpreter error."""
-
-    def __init__(self, message, ast=None):
-        super().__init__(message)
-        self.parseinfo = ast.parseinfo if ast is not None else None
 
 
 class EvalNode:
@@ -391,10 +376,6 @@ class EvalCoalesce(EvalNode):
     __slots__ = ('args',)
 
     def __init__(self, args):
-        for arg in args:
-            if arg.dtype != args[0].dtype:
-                dtypes = ', '.join(arg.dtype.__name__ for arg in args)
-                raise CompilationError(f'coalesce() function arguments must have uniform type, found: {dtypes}')
         super().__init__(args[0].dtype)
         self.args = args
 
@@ -529,197 +510,6 @@ class SubqueryTable(tables.Table):
         return iter(rows)
 
 
-def compile_expression(expr, environ):
-    """Bind an expression to its execution context.
-
-    Args:
-      expr: The root node of an expression.
-      environ: An CompilationEnvironment instance.
-    Returns:
-      The root node of a bound expression.
-    """
-
-    if expr is None:
-        return None
-
-    if isinstance(expr, ast.Column):
-        column = environ.columns.get(expr.name)
-        if column is not None:
-            return column
-        raise CompilationError(f'column "{expr.name}" does not exist', expr)
-
-    if isinstance(expr, ast.Or):
-        return EvalOr([compile_expression(arg, environ) for arg in expr.args])
-
-    if isinstance(expr, ast.And):
-        return EvalAnd([compile_expression(arg, environ) for arg in expr.args])
-
-    if isinstance(expr, ast.Function):
-        operands = [compile_expression(operand, environ) for operand in expr.operands]
-        if expr.fname == 'coalesce':
-            # coalesce() is parsed like a function call but it does
-            # not really fit our model for function evaluation,
-            # therefore it gets special threatment here.
-            return EvalCoalesce(operands)
-        function = types.function_lookup(FUNCTIONS, expr.fname, operands)
-        if function is None:
-            sig = '{}({})'.format(expr.fname, ', '.join(f'{operand.dtype.__name__.lower()}' for operand in operands))
-            raise CompilationError(f'no function matches "{sig}" name and argument types', expr)
-        function = function(operands)
-        # Constants folding.
-        if all(isinstance(operand, EvalConstant) for operand in operands) and function.pure:
-            return EvalConstant(function(None), function.dtype)
-        return function
-
-    if isinstance(expr, ast.Subscript):
-        operand = compile_expression(expr.operand, environ)
-        if issubclass(operand.dtype, dict):
-            return EvalGetItem(operand, expr.key)
-        raise CompilationError('column type is not subscriptable', expr)
-
-    if isinstance(expr, ast.Attribute):
-        operand = compile_expression(expr.operand, environ)
-        dtype = types.ALIASES.get(operand.dtype, operand.dtype)
-        if issubclass(dtype, types.Structure):
-            getter = dtype.columns.get(expr.name)
-            if getter is None:
-                raise CompilationError(f'structured type has no attribute "{expr.name}"', expr)
-            return EvalGetter(operand, getter, getter.dtype)
-        raise CompilationError('column type is not structured', expr)
-
-    if isinstance(expr, ast.UnaryOp):
-        operand = compile_expression(expr.operand, environ)
-        function = types.function_lookup(OPERATORS, type(expr), [operand])
-        if function is None:
-            raise CompilationError(
-                f'operator "{type(expr).__name__.lower()}({operand.dtype.__name__})" not supported', expr)
-        function = function(operand)
-        # Constants folding.
-        if isinstance(operand, EvalConstant):
-            return EvalConstant(function(None), function.dtype)
-        return function
-
-    if isinstance(expr, ast.Between):
-        operand = compile_expression(expr.operand, environ)
-        lower = compile_expression(expr.lower, environ)
-        upper = compile_expression(expr.upper, environ)
-        intypes = [operand.dtype, lower.dtype, upper.dtype]
-        for candidate in OPERATORS[type(expr)]:
-            if candidate.__intypes__ == intypes:
-                func = candidate(operand, lower, upper)
-                return func
-        raise CompilationError(
-            f'operator "{types.name(operand.dtype)} BETWEEN {types.name(lower.dtype)} '
-            f'AND {types.name(upper.dtype)}" not supported', expr)
-
-    if isinstance(expr, ast.BinaryOp):
-        left = compile_expression(expr.left, environ)
-        right = compile_expression(expr.right, environ)
-
-        candidates = OPERATORS[type(expr)]
-        while True:
-            intypes = [left.dtype, right.dtype]
-            for op in candidates:
-                if op.__intypes__ == intypes:
-                    function = op(left, right)
-                    # Constants folding.
-                    if isinstance(left, EvalConstant) and isinstance(right, EvalConstant):
-                        return EvalConstant(function(None), function.dtype)
-                    return function
-
-            # Implement type inference when one of the operands is not strongly typed.
-            if left.dtype is object and right.dtype is not object:
-                target = right.dtype
-                if target is int:
-                    # The Beancount parser does not emit int typed
-                    # values, thus casting to int is only going to
-                    # loose information. Promote to decimal.
-                    target = Decimal
-                name = types.MAP.get(target)
-                if name is None:
-                    break
-                left = types.function_lookup(FUNCTIONS, name, [left])([left])
-                continue
-            if right.dtype is object and left.dtype is not object:
-                target = left.dtype
-                if target is int:
-                    # The Beancount parser does not emit int typed
-                    # values, thus casting to int is only going to
-                    # loose information. Promote to decimal.
-                    target = Decimal
-                name = types.MAP.get(target)
-                if name is None:
-                    break
-                right = types.function_lookup(FUNCTIONS, name, [right])([right])
-                continue
-
-            # Failure.
-            break
-
-        raise CompilationError(
-            f'operator "{type(expr).__name__.lower()}('
-            f'{left.dtype.__name__}, {right.dtype.__name__})" not supported', expr)
-
-    if isinstance(expr, ast.Constant):
-        return EvalConstant(expr.value)
-
-    if isinstance(expr, ast.Asterisk):
-        return EvalConstant(None, dtype=types.Asterisk)
-
-    raise NotImplementedError
-
-
-def get_columns_and_aggregates(node):
-    """Find the columns and aggregate nodes below this tree.
-
-    All nodes under aggregate nodes are ignored.
-
-    Args:
-      node: An instance of EvalNode.
-    Returns:
-      A pair of (columns, aggregates), both of which are lists of EvalNode instances.
-        columns: The list of all columns accessed not under an aggregate node.
-        aggregates: The list of all aggregate nodes.
-    """
-    columns = []
-    aggregates = []
-    _get_columns_and_aggregates(node, columns, aggregates)
-    return columns, aggregates
-
-
-def _get_columns_and_aggregates(node, columns, aggregates):
-    """Walk down a tree of nodes and fetch the column accessors and aggregates.
-
-    This function ignores all nodes under aggregate nodes.
-
-    Args:
-      node: An instance of EvalNode.
-      columns: An accumulator for columns found so far.
-      aggregate: An accumulator for aggregate notes found so far.
-    """
-    if isinstance(node, EvalAggregator):
-        aggregates.append(node)
-    elif isinstance(node, EvalColumn):
-        columns.append(node)
-    else:
-        for child in node.childnodes():
-            _get_columns_and_aggregates(child, columns, aggregates)
-
-
-def is_aggregate(node):
-    """Return true if the node is an aggregate.
-
-    Args:
-      node: An instance of EvalNode.
-    Returns:
-      A boolean.
-    """
-    # Note: We could be a tiny bit more efficient here, but it doesn't matter
-    # much. Performance of the query compilation matters very little overall.
-    _, aggregates = get_columns_and_aggregates(node)
-    return bool(aggregates)
-
-
 # A compiled target.
 #
 # Attributes:
@@ -730,296 +520,6 @@ def is_aggregate(node):
 EvalTarget = collections.namedtuple('EvalTarget', 'c_expr name is_aggregate')
 
 
-def get_target_name(target):
-    """Compute the target name.
-
-    This uses the same algorithm used by SQLite. If the target has an
-    AS clause assigning it a name, that will be the name used. If the
-    target refers directly to a column, then the target name is the
-    column name. Otherwise use the expression text.
-
-    """
-    if target.name is not None:
-        return target.name
-    if isinstance(target.expression, ast.Column):
-        return target.expression.name
-    return target.expression.text.strip()
-
-
-def compile_targets(targets, environ):
-    """Compile the targets and check for their validity. Process wildcard.
-
-    Args:
-      targets: A list of target expressions from the parser.
-      environ: A compilation context for the targets.
-    Returns:
-      A list of compiled target expressions with resolved names.
-    """
-    # Bind the targets expressions to the execution context.
-    if isinstance(targets, ast.Asterisk):
-        # Insert the full list of available columns.
-        targets = [ast.Target(ast.Column(name), None)
-                   for name in environ.wildcard_columns]
-
-    # Compile targets.
-    c_targets = []
-    for target in targets:
-        c_expr = compile_expression(target.expression, environ)
-        name = get_target_name(target)
-        c_targets.append(EvalTarget(c_expr, name, is_aggregate(c_expr)))
-
-        columns, aggregates = get_columns_and_aggregates(c_expr)
-
-        # Check for mixed aggregates and non-aggregates.
-        if columns and aggregates:
-            raise CompilationError('mixed aggregates and non-aggregates are not allowed')
-
-        # Check for aggregates of aggregates.
-        for aggregate in aggregates:
-            for child in aggregate.childnodes():
-                if is_aggregate(child):
-                    raise CompilationError('aggregates of aggregates are not allowed')
-
-    return c_targets
-
-
-def compile_group_by(group_by, c_targets, environ):
-    """Process a group-by clause.
-
-    Args:
-      group_by: A GroupBy instance as provided by the parser.
-      c_targets: A list of compiled target expressions.
-      environ: A compilation context to be used to evaluate GROUP BY expressions.
-    Returns:
-      A tuple of
-       new_targets: A list of new compiled target nodes.
-       group_indexes: If the query is an aggregate query, a list of integer
-         indexes to be used for processing grouping. Note that this list may be
-         empty (in the case of targets with only aggregates). On the other hand,
-         if this is not an aggregated query, this is set to None. So do
-         distinguish the empty list vs. None.
-    """
-    new_targets = copy.copy(c_targets)
-    c_target_expressions = [c_target.c_expr for c_target in c_targets]
-
-    group_indexes = []
-    having_index = None
-
-    if group_by:
-        assert group_by.columns, "Internal error with GROUP-BY parsing"
-
-        # Compile group-by expressions and resolve them to their targets if
-        # possible. A GROUP-BY column may be one of the following:
-        #
-        # * A reference to a target by name.
-        # * A reference to a target by index (starting at one).
-        # * A new, non-aggregate expression.
-        #
-        # References by name are converted to indexes. New expressions are
-        # inserted into the list of targets as invisible targets.
-        targets_name_map = {target.name: index for index, target in enumerate(c_targets)}
-        for column in group_by.columns:
-            index = None
-
-            # Process target references by index.
-            if isinstance(column, int):
-                index = column - 1
-                if not 0 <= index < len(c_targets):
-                    raise CompilationError(f'invalid GROUP-BY column index {column}')
-
-            else:
-                # Process target references by name. These will be parsed as
-                # simple Column expressions. If they refer to a target name, we
-                # resolve them.
-                if isinstance(column, ast.Column):
-                    name = column.name
-                    index = targets_name_map.get(name, None)
-
-                # Otherwise we compile the expression and add it to the list of
-                # targets to evaluate and index into that new target.
-                if index is None:
-                    c_expr = compile_expression(column, environ)
-
-                    # Check if the new expression is an aggregate.
-                    aggregate = is_aggregate(c_expr)
-                    if aggregate:
-                        raise CompilationError(f'GROUP-BY expressions may not be aggregates: "{column}"')
-
-                    # Attempt to reconcile the expression with one of the existing
-                    # target expressions.
-                    try:
-                        index = c_target_expressions.index(c_expr)
-                    except ValueError:
-                        # Add the new target. 'None' for the target name implies it
-                        # should be invisible, not to be rendered.
-                        index = len(new_targets)
-                        new_targets.append(EvalTarget(c_expr, None, aggregate))
-                        c_target_expressions.append(c_expr)
-
-            assert index is not None, "Internal error, could not index group-by reference."
-            group_indexes.append(index)
-
-            # Check that the group-by column references a non-aggregate.
-            c_expr = new_targets[index].c_expr
-            if is_aggregate(c_expr):
-                raise CompilationError(f'GROUP-BY expressions may not reference aggregates: "{column}"')
-
-            # Check that the group-by column has a supported hashable type.
-            if not issubclass(c_expr.dtype, collections.abc.Hashable):
-                raise CompilationError(f'GROUP-BY a non-hashable type is not supported: "{column}"')
-
-        # Compile HAVING clause.
-        if group_by.having is not None:
-            c_expr = compile_expression(group_by.having, environ)
-            if not is_aggregate(c_expr):
-                raise CompilationError('the HAVING clause must be an aggregate expression')
-            having_index = len(new_targets)
-            new_targets.append(EvalTarget(c_expr, None, True))
-            c_target_expressions.append(c_expr)
-
-    else:
-        # If it does not have a GROUP-BY clause...
-        aggregate_bools = [c_target.is_aggregate for c_target in c_targets]
-        if any(aggregate_bools):
-            # If the query is an aggregate query, check that all the targets are
-            # aggregates.
-            if all(aggregate_bools):
-                # FIXME: shold we really be checking for the empty
-                # list or is checking for a false value enough?
-                assert group_indexes == []
-            elif SUPPORT_IMPLICIT_GROUPBY:
-                # If some of the targets aren't aggregates, automatically infer
-                # that they are to be implicit group by targets. This makes for
-                # a much more convenient syntax for our lightweight SQL, where
-                # grouping is optional.
-                group_indexes = [
-                    index for index, c_target in enumerate(c_targets)
-                    if not c_target.is_aggregate]
-            else:
-                raise CompilationError('aggregate query without a GROUP-BY should have only aggregates')
-        else:
-            # This is not an aggregate query; don't set group_indexes to
-            # anything useful, we won't need it.
-            group_indexes = None
-
-    return new_targets[len(c_targets):], group_indexes, having_index
-
-
-def compile_order_by(order_by, c_targets, environ):
-    """Process an order-by clause.
-
-    Args:
-      order_by: A OrderBy instance as provided by the parser.
-      c_targets: A list of compiled target expressions.
-      environ: A compilation context to be used to evaluate ORDER BY expressions.
-    Returns:
-      A tuple of
-       new_targets: A list of new compiled target nodes.
-       order_spec: A list of (integer indexes, sort order) tuples.
-    """
-    if not order_by:
-        return [], None
-
-    new_targets = copy.copy(c_targets)
-    c_target_expressions = [c_target.c_expr for c_target in c_targets]
-    order_spec = []
-
-    # Compile order-by expressions and resolve them to their targets if
-    # possible. A ORDER-BY column may be one of the following:
-    #
-    # * A reference to a target by name.
-    # * A reference to a target by index (starting at one).
-    # * A new expression, aggregate or not.
-    #
-    # References by name are converted to indexes. New expressions are
-    # inserted into the list of targets as invisible targets.
-    targets_name_map = {target.name: index for index, target in enumerate(c_targets)}
-    for spec in order_by:
-        column = spec.column
-        descending = spec.ordering
-        index = None
-
-        # Process target references by index.
-        if isinstance(column, int):
-            index = column - 1
-            if not 0 <= index < len(c_targets):
-                raise CompilationError(f'invalid ORDER-BY column index {column}')
-
-        else:
-            # Process target references by name. These will be parsed as
-            # simple Column expressions. If they refer to a target name, we
-            # resolve them.
-            if isinstance(column, ast.Column):
-                name = column.name
-                index = targets_name_map.get(name, None)
-
-            # Otherwise we compile the expression and add it to the list of
-            # targets to evaluate and index into that new target.
-            if index is None:
-                c_expr = compile_expression(column, environ)
-
-                # Attempt to reconcile the expression with one of the existing
-                # target expressions.
-                try:
-                    index = c_target_expressions.index(c_expr)
-                except ValueError:
-                    # Add the new target. 'None' for the target name implies it
-                    # should be invisible, not to be rendered.
-                    index = len(new_targets)
-                    new_targets.append(EvalTarget(c_expr, None, is_aggregate(c_expr)))
-                    c_target_expressions.append(c_expr)
-
-        assert index is not None, "Internal error, could not index order-by reference."
-        order_spec.append((index, descending))
-
-    return new_targets[len(c_targets):], order_spec
-
-
-def compile_pivot_by(pivot_by, targets, group_indexes):
-    """Compiles a PIVOT BY clause.
-
-    Resolve and validate columns references in the PIVOT BY clause.
-    The PIVOT BY clause accepts two name od index references to
-    columns in the SELECT targets list. The second columns should be a
-    GROUP BY column so that the values of the pivot column are unique.
-
-    """
-    if pivot_by is None:
-        return None
-
-    indexes = []
-    names = {target.name: index for index, target in enumerate(targets)}
-
-    for column in pivot_by.columns:
-
-        # Process target references by index.
-        if isinstance(column, int):
-            index = column - 1
-            if not 0 <= index < len(targets):
-                raise CompilationError(f'invalid PIVOT BY column index {column}')
-            indexes.append(index)
-            continue
-
-        # Process target references by name.
-        if isinstance(column, ast.Column):
-            index = names.get(column.name, None)
-            if index is None:
-                raise CompilationError(f'PIVOT BY column {column!r} is not in the targets list')
-            indexes.append(index)
-            continue
-
-        # Not reached.
-        raise RuntimeError
-
-    # Sanity checks.
-    if indexes[0] == indexes[1]:
-        raise CompilationError('the two PIVOT BY columns cannot be the same column')
-    if indexes[1] not in group_indexes:
-        raise CompilationError('the second PIVOT BY column must be a GROUP BY column')
-
-    return indexes
-
-
 # A compile FROM clause.
 #
 # Attributes:
@@ -1027,33 +527,6 @@ def compile_pivot_by(pivot_by, targets, group_indexes):
 #   close:
 #   clear: See ast.From attributes description.
 EvalFrom = collections.namedtuple('EvalFrom', 'open close clear')
-
-def compile_from(from_clause, environ):
-    """Compiled a FROM clause as provided by the parser, in the given environment.
-
-    Args:
-      from_clause: An instance of ast.From.
-      environ: Environment.
-    Returns:
-      An tuple of EvalFrom and EvalNode instances.
-    """
-    if from_clause is None:
-        return None, None
-
-    c_expression = compile_expression(from_clause.expression, environ)
-
-    # Check that the FROM clause does not contain aggregates.
-    if c_expression is not None and is_aggregate(c_expression):
-        raise CompilationError('aggregates are not allowed in FROM clause')
-
-    if (isinstance(from_clause.open, datetime.date) and
-        isinstance(from_clause.close, datetime.date) and
-        from_clause.open > from_clause.close):
-        raise CompilationError('CLOSE date must follow OPEN date')
-
-    c_from = EvalFrom(from_clause.open, from_clause.close, from_clause.clear)
-
-    return c_from, c_expression
 
 
 # A compiled query, ready for execution.
@@ -1088,206 +561,9 @@ EvalQuery = collections.namedtuple('EvalQuery', ('table c_targets c_where '
 EvalPivot = collections.namedtuple('EvalPivot', 'query pivots')
 
 
-def compile_select(context, select):
-    """Prepare an AST for a Select statement into a very rudimentary execution tree.
-    The execution tree mostly looks much like an AST, but with some nodes
-    replaced with knowledge specific to an execution context and eventually some
-    basic optimizations.
-
-    Args:
-      select: An instance of ast.Select.
-    Returns:
-      An instance of EvalQuery, ready to be executed.
-    """
-
-    # Compile the FROM clause.
-    if select.from_clause is None:
-        # Implicit table reference.
-        table = context.tables.get('postings')
-        c_from_expr = None
-
-    elif isinstance(select.from_clause, ast.Select):
-        # Subquery.
-        table = SubqueryTable(compile_select(context, select.from_clause))
-        c_from_expr = None
-
-    elif isinstance(select.from_clause, ast.Table):
-        # Table reference.
-        table = context.tables.get(select.from_clause.name)
-        if table is None:
-            raise CompilationError(f'table "{select.from_clause.name}" does not exist', select.from_clause)
-        c_from_expr = None
-
-    else:
-        # FROM expression. FROM expressions filter transactions thus
-        # use the columns definitions for the ``entries`` table.
-        table = context.tables.get('postings')
-        c_from, c_from_expr = compile_from(select.from_clause, context.tables.get('entries'))
-        if c_from is not None:
-            table = table.update(**c_from._asdict())
-
-    # Compile the targets.
-    c_targets = compile_targets(select.targets, table)
-
-    # Bind the WHERE expression to the execution environment.
-    c_where = compile_expression(select.where_clause, table)
-
-    # Check that the FROM clause does not contain aggregates. This
-    # should never trigger if the compilation environment does not
-    # contain any aggregate.
-    if c_where is not None and is_aggregate(c_where):
-        raise CompilationError('aggregates are not allowed in WHERE clause')
-
-    # Combine FROM and WHERE clauses
-    if c_from_expr is not None:
-        c_where = c_from_expr if c_where is None else EvalAnd([c_from_expr, c_where])
-
-    # Process the GROUP-BY clause.
-    new_targets, group_indexes, having_index = compile_group_by(select.group_by,
-                                                                c_targets,
-                                                                table)
-    c_targets.extend(new_targets)
-
-    # Process the ORDER-BY clause.
-    new_targets, order_spec = compile_order_by(select.order_by,
-                                               c_targets,
-                                               table)
-    c_targets.extend(new_targets)
-
-    # If this is an aggregate query (it groups, see list of indexes), check that
-    # the set of non-aggregates match exactly the group indexes. This should
-    # always be the case at this point, because we have added all the necessary
-    # targets to the list of group-by expressions and should have resolved all
-    # the indexes.
-    if group_indexes is not None:
-        non_aggregate_indexes = {index for index, c_target in enumerate(c_targets)
-                                 if not c_target.is_aggregate}
-        if non_aggregate_indexes != set(group_indexes):
-            missing_names = ['"{}"'.format(c_targets[index].name)
-                             for index in non_aggregate_indexes - set(group_indexes)]
-            raise CompilationError(
-                'all non-aggregates must be covered by GROUP-BY clause in aggregate query: '
-                'the following targets are missing: {}'.format(','.join(missing_names)))
-
-    query = EvalQuery(table,
-                      c_targets,
-                      c_where,
-                      group_indexes,
-                      having_index,
-                      order_spec,
-                      select.limit,
-                      select.distinct)
-
-    pivots = compile_pivot_by(select.pivot_by, c_targets, group_indexes)
-    if pivots:
-        return EvalPivot(query, pivots)
-
-    return query
-
-
-def transform_journal(journal):
-    """Translate a Journal entry into an uncompiled Select statement.
-
-    Args:
-      journal: An instance of a Journal object.
-    Returns:
-      An instance of an uncompiled Select object.
-    """
-    cooked_select = parser.parse("""
-
-        SELECT
-           date,
-           flag,
-           MAXWIDTH(payee, 48),
-           MAXWIDTH(narration, 80),
-           account,
-           {summary_func}(position),
-           {summary_func}(balance)
-        {where}
-
-    """.format(where=('WHERE account ~ "{}"'.format(journal.account)
-                      if journal.account
-                      else ''),
-               summary_func=journal.summary_func or ''))
-
-    return ast.Select(cooked_select.targets,
-                      journal.from_clause,
-                      cooked_select.where_clause,
-                      None, None, None, None, None)
-
-
-def transform_balances(balances):
-    """Translate a Balances entry into an uncompiled Select statement.
-
-    Args:
-      balances: An instance of a Balance object.
-    Returns:
-      An instance of an uncompiled Select object.
-    """
-    ## FIXME: Change the aggregation rules to allow GROUP-BY not to include the
-    ## non-aggregate ORDER-BY columns, so we could just GROUP-BY accounts here
-    ## instead of having to include the sort-key. I think it should be fine if
-    ## the first or last sort-order value gets used, because it would simplify
-    ## the input statement.
-
-    cooked_select = parser.parse("""
-
-      SELECT account, SUM({}(position))
-      GROUP BY account, ACCOUNT_SORTKEY(account)
-      ORDER BY ACCOUNT_SORTKEY(account)
-
-    """.format(balances.summary_func or ""))
-
-    return ast.Select(cooked_select.targets,
-                      balances.from_clause,
-                      balances.where_clause,
-                      cooked_select.group_by,
-                      cooked_select.order_by,
-                      None, None, None)
-
-
 # A compiled print statement, ready for execution.
 #
 # Attributes:
 #   table: Table to print
 #   where: Filtering expression, EvalNode instance.
 EvalPrint = collections.namedtuple('EvalPrint', 'table where')
-
-def compile_print(context, print_stmt):
-    """Compile a Print statement.
-
-    Args:
-      statement: An instance of ast.Print.
-    Returns:
-      An instance of EvalPrint, ready to be executed.
-    """
-    # Compile FROM clause.
-    table = context.tables.get('entries')
-    c_from, expr = compile_from(print_stmt.from_clause, table)
-    if c_from is not None:
-        table = table.update(**c_from._asdict())
-    return EvalPrint(table, expr)
-
-
-def compile(context, statement):
-    """Prepare an AST any of the statement into an executable statement.
-
-    Args:
-      statement: An instance of the parser's Select, Balances, Journal or Print.
-    Returns:
-      An instance of EvalQuery or EvalPrint, ready to be executed.
-    Raises:
-      CompilationError: If the statement cannot be compiled, or is not one of the
-        supported statements.
-    """
-    if isinstance(statement, ast.Balances):
-        statement = transform_balances(statement)
-    elif isinstance(statement, ast.Journal):
-        statement = transform_journal(statement)
-
-    if isinstance(statement, ast.Select):
-        return compile_select(context, statement)
-    if isinstance(statement, ast.Print):
-        return compile_print(context, statement)
-
-    raise NotImplementedError
